+++ pyramid/tests/pkgs/notfoundview/__init__.py
@@ -1,4 +1,4 @@
-from pyramid.view import notfound_view_config, view_config
+from pyramid.view import notfound_view_config, view_config, UseSubrequest
 from pyramid.response import Response
 
 @notfound_view_config(route_name='foo', append_slash=True)
@@ -21,10 +21,20 @@ def bar(request):
 def foo2(request):
     return Response('OK foo2')
 
+@notfound_view_config(route_name='wiz', append_slash=UseSubrequest)
+def wiz_notfound(request): # pragma: no cover
+    return Response('wiz_notfound')
+
+@view_config(route_name='wiz2')
+def wiz2(request):
+    return Response('OK wiz2')
+
 def includeme(config):
     config.add_route('foo', '/foo')
     config.add_route('foo2', '/foo/')
     config.add_route('bar', '/bar/')
     config.add_route('baz', '/baz')
+    config.add_route('wiz', '/wiz')
+    config.add_route('wiz2', '/wiz/')
     config.scan('pyramid.tests.pkgs.notfoundview')